modules/person-grpc-service/requirements.txt
grpcio-tools==1.71.*
SQLAlchemy==2.0.*
marshmallow==3.26.*
marshmallow-sqlalchemy==1.4.*
psycopg2-binary==2.9.*
environs==14.1.*